% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{adjustTransitivity}
\alias{adjustTransitivity}
\title{adjustTransitivity}
\usage{
adjustTransitivity(graph, target_transitivity = 0.5, max_iter = 1000)
}
\arguments{
\item{graph}{Igraph object}

\item{target_transitivity}{target_transitivity measure for the simulated graph. (Default is .5)}

\item{max_iter}{the maximum number of iterations to try to achieve the target transitivity.}
}
\value{
random graph with similar in and out degree
}
\description{
adjustTransitivity iteratively rewires the graph to achieve the desired transitivity.
}
