% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DREAM_data.R
\name{aggregateExpression}
\alias{aggregateExpression}
\title{aggregateExpression}
\usage{
aggregateExpression(training_directory, perturbation = NULL)
}
\arguments{
\item{training_directory}{Path to the input directory of training data}

\item{perturbation}{The in silico perturbation condition. Default is NULL, in which case all perturbation conditions are combined. Other options include "wildtype", "knockdown", "knockouts", and "multifactorial"}
}
\value{
A list containing the following components:
\item{expression}{Aggregate expression dataset}
\item{metadata}{Aggregate metadata labels}
}
\description{
This function reads in the simulated expression profiles from the DREAM4 challenge. Depending on the perturbation parameter, it will read in a single data type or all available.
}
