% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DREAM_data.R
\name{aggregateNetworks}
\alias{aggregateNetworks}
\title{aggregateNetworks}
\usage{
aggregateNetworks(gold_standards_directory)
}
\arguments{
\item{gold_standards_directory}{Directory where the gold standard networks exist. Gold standard networks are edgelists with the format (G1    G2    1)}
}
\value{
network Aggregated network
}
\description{
This function reads in the gene regulatory networks used to create the simulated expression data. .
}
