% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_processsing_functions.R
\name{alignNetwork}
\alias{alignNetwork}
\title{alignNetwork}
\usage{
alignNetwork(
  expression_file,
  adjacency_file,
  probability_threshold = 1,
  tissue,
  correlation_threshold = 0.05
)
}
\arguments{
\item{expression_file}{GTEX expression file, downloaded from GRAND. (For example, see https://grand.networkmedicine.org/tissues/Breast_tissue/)}

\item{adjacency_file}{Network adjacency downloaded from GRAND. (For exampple, see https://grand.networkmedicine.org/tissues/Breast_tissue/)}

\item{probability_threshold}{Number to threshold the filtering of the graph edge weights. (Default =1)}

\item{tissue}{String of the tissue type for the GRN and expression data}

\item{correlation_threshold}{Number representing the threshold for significant correlations for signing the network. (Default = .05)}
}
\description{
alignNetwork is a function for reading in, filtering, and otherwise processing GRNs downloaded from GRAND for use in GRACKLE.
}
