% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ensemblToHGNC}
\alias{ensemblToHGNC}
\title{entrezToHGNC}
\usage{
ensemblToHGNC(ensembl_ids)
}
\arguments{
\item{ensembl_ids}{Character vector of entrez ID's to convert to HGNC gene ID's}
}
\value{
hgnc_ids   Character vector of HGNC ID's
}
\description{
entrezToHGNC is a function for converting entrez ID's to HGNC ID's
}
