% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euclideanDist}
\alias{euclideanDist}
\title{Calculate Euclidean Distance Matrix}
\usage{
euclideanDist(X)
}
\arguments{
\item{X}{A numeric matrix where rows represent observations and columns represent variables.}
}
\value{
A numeric matrix representing the pairwise Euclidean distances between rows of the input matrix.
}
\description{
Calculate Euclidean Distance Matrix
}
