% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{evaluationWrapper}
\alias{evaluationWrapper}
\title{evaluationWrapper}
\usage{
evaluationWrapper(
  test_expression,
  test_metadata,
  H_train,
  k,
  clusters,
  aligned_clusters
)
}
\arguments{
\item{test_expression}{Data.frame of test expression data}

\item{test_metadata}{Data.frame of the test metadata}

\item{H_train}{Matrix of trained H matrix}

\item{k}{Numeric value representing the number of groups}

\item{clusters}{igraph communities object with cluster membership}

\item{aligned_clusters}{clusters aligned with metadata assignments}
}
\description{
evaluationWrapper is function for finding the alignment of results
}
