% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{geneLoadingsEvaluation}
\alias{geneLoadingsEvaluation}
\title{geneLoadingsEvaluation}
\usage{
geneLoadingsEvaluation(H_train, module_size = 10)
}
\arguments{
\item{H_train}{Matrix of trained H matrix}

\item{module_size}{Size of consecutive modules for DREAM simulations.}
}
\description{
geneLoadingsEvaluation is a function for identifying the top loading genes on each latent variable and aligning them with input data
}
