% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRACKLE.R
\name{GRACKLE}
\alias{GRACKLE}
\title{GRACKLE}
\usage{
GRACKLE(
  Y,
  net_similarity,
  patient_similarity,
  diff_threshold = 1e-04,
  lambda_1 = 0.5,
  lambda_2 = 0.5,
  k = 5,
  verbose = FALSE,
  beta = 0
)
}
\arguments{
\item{Y}{Input gene expression data}

\item{net_similarity}{Similarity matrix based on GRN}

\item{patient_similarity}{Similarity matrix based on patient metadata}

\item{diff_threshold}{A numeric value representing the threshold for the difference in W and H matrices between iterations. Default is 1e-4}

\item{lambda_1}{A numeric value representing the first regularization parameter. Default is 0.5.}

\item{lambda_2}{A numeric value representing the second regularization parameter. Default is 0.5.}

\item{k}{An integer specifying the number of clusters or components. Default is 5.}

\item{verbose}{A boolean value as to whether to print loss error or not. Default is FALSE.}

\item{beta}{A numeric value representing the degree of l2 regularization. Default is 0.}
}
\value{
A list containing the following components:
\item{residual}{A matrix representing the residuals, calculated as \code{Y - W \%*\% H}.}
\item{H}{The matrix \code{H} obtained from the NMF calculation.}
\item{W}{The matrix \code{W} obtained from the NMF calculation.}
\item{error}{A vector containing the reconstruction error at each iteration.}
\item{H_diff}{A vector containing the differences in \code{H} between iterations.}
\item{W_diff}{A vector containing the differences in \code{W} between iterations.}
\item{pat_sim_error_vec}{A vector containing the patient similarity error at each iteration.}
\item{grn_error_vec}{A vector containing the gene regulatory network error at each iteration.}
}
\description{
Function for running Graph Regularization Across Contextual KnowLedgE
}
