% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DREAM_data.R
\name{imputeExpression}
\alias{imputeExpression}
\title{imputeExpression}
\usage{
imputeExpression(expression_matrix)
}
\arguments{
\item{expression_matrix}{Data.frame of simulated gene expression data.}
}
\value{
expression_matrix Imputed matrix
}
\description{
This function imputes missing values in the gene expression matrix randomly.
}
