% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{increaseTransitivity}
\alias{increaseTransitivity}
\title{increaseTransitivity}
\usage{
increaseTransitivity(graph)
}
\arguments{
\item{graph}{Igraph object}
}
\value{
random graph with similar in and out degree
}
\description{
Function for increasing the transitivity of the graph based on
}
