% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{makeDirected}
\alias{makeDirected}
\title{makeDirected}
\usage{
makeDirected(expression, network)
}
\arguments{
\item{expression}{Data Frame of gene expression data}

\item{network}{Data Frame of the network edgelist with column names 'from', 'to', 'weight'.}
}
\value{
Dataframe of network edgelist with direction (1 = activation, -1 = suppression)
}
\description{
makedDirected is a function for  calculating the pairwise correlations coefficients between genes and assigning direction to the network based on correlation.
}
