% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{modulePermutations}
\alias{modulePermutations}
\title{modulePermutations}
\usage{
modulePermutations(g, membership, module_ids)
}
\arguments{
\item{g}{igraph graph object}

\item{membership}{module membership based on graph clustering.}

\item{module_ids}{Numeric vector for the membership ids to permute.}
}
\value{
permuted graph with the original degree distribution.
}
\description{
modulePermutations is a function for creating permuted graphs based on permuting specific modules within the graph
}
