% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(matrix)
}
\arguments{
\item{matrix}{input matrix to normalize}
}
\value{
normalized matrix
}
\description{
normalize is a function for scaling matrices
}
