% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{parallelSimulateExpression}
\alias{parallelSimulateExpression}
\title{parallelSimulateExpression}
\usage{
parallelSimulateExpression(
  g,
  max_expression = 2000,
  num_samples = 5,
  iterations = 3,
  rc = NULL,
  select_nodes = NULL,
  perturbation_constant = 1.5,
  noise_constant = 0.3
)
}
\arguments{
\item{g}{igraph graph object}

\item{max_expression}{Numeric value for the maximum_gene expression. (Default is 2000)}

\item{num_samples}{Numeric values for the number of samples to simulate. (Default is 5)}

\item{iterations}{numeric value to pass to simulation function. (Default is 3)}

\item{select_nodes}{Numeric vector of specific nodes to perturb}

\item{perturbation_constant}{Numberic value to increase initial values of perturbed nodes by. (Default is 1.5)}

\item{noise_constant}{Numeric value representing how much of the rest of the matrix to add noise to.}
}
\value{
data frame of simulated gene expression data
}
\description{
parallelizes the simulate expression function to simulate the gene expression for multiple samples in the same subgroup
}
