% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DREAM_data.R
\name{permuteNetwork}
\alias{permuteNetwork}
\title{permuteNetwork}
\usage{
permuteNetwork(network, replace_percentage = 0.1)
}
\arguments{
\item{network}{Matrix for the adjacency to permute}
}
\value{
modified_matrix Matrix of the permuted network
}
\description{
permuteNetwork is a function for permuting the network to test as a negative control
}
