% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{plotSimulatedExpression}
\alias{plotSimulatedExpression}
\title{plotSimulatedExpression}
\usage{
plotSimulatedExpression(expression, metadata)
}
\arguments{
\item{expression}{simulated gene expression data. sample X genes}

\item{metadata}{simulated metadata. sample by column}
}
\description{
plots heatmap of simulated expression
}
