% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DREAM_data.R
\name{preserveCorrelation}
\alias{preserveCorrelation}
\title{preserveCorrelation}
\usage{
preserveCorrelation(original, noisy)
}
\arguments{
\item{original}{A matrix of the original gene expression values. NA values set to 0.}

\item{noisy}{A matrix of the noisy gene expression values, including NA values.}
}
\value{
expression_matrix Imputed matrix
}
\description{
preserveCorrelation is a function that uses pca to ensure that noisy data maintains the original correlation structure
}
