% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{project_W}
\alias{project_W}
\title{project_W}
\usage{
project_W(test_data, H_train, k)
}
\arguments{
\item{H_train}{Matrix of trained H matrix}

\item{k}{Numeric value representing the number of groups}

\item{test_expression}{Data.frame of test expression data}
}
\value{
W_test Matrix of project W values for the test data
}
\description{
Function for predicting categorical labels in test data
}
