% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{randomNetwork}
\alias{randomNetwork}
\title{randomNetwork}
\usage{
randomNetwork(n = 100, plot = FALSE)
}
\arguments{
\item{n}{Numeric value indicating the size of the graph. (Default is 100)}
}
\value{
random graph with the defined number of edges
}
\description{
randomNetwork applies igraph functions for creating a random network
}
