% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{randomQuantileVector}
\alias{randomQuantileVector}
\title{randomQantileVector}
\usage{
randomQuantileVector(n, values)
}
\arguments{
\item{n}{Numeric value for the length of random vector}

\item{values}{Numeric vector of length 5 with reported quantile values}
}
\value{
random quantile vector of length n
}
\description{
randomQunatileVector calculates a random vector based on an input quantile vector for quantiles (0,.25,.5,.75,1)
}
