% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarks.R
\name{runGRNMF}
\alias{runGRNMF}
\title{runGRNMF}
\usage{
runGRNMF(expression_matrix, k = 5, seed = 42, max_iter = 200, alpha = 0.1)
}
\arguments{
\item{expression_matrix}{Matrix of gene expression data (samples as rows, genes as columns)}

\item{k}{A numeric variable representing the number of latent variables (aka, rank). Default is 5.}

\item{seed}{Specifies the starting point or seeding method. See the NMF package for details.  (Default is 42.)}

\item{max_iter}{Numeric for the maximum number of iterations. (Default is 200)}

\item{alpha}{Numeric value representing the effect of graph regularization on the model. (Default is .1)}

\item{nearest_neighbors}{A numeric value representing the number of nearest neighbors. (Default is 5. )}
}
\description{
runGRNMF implements a version of graph regularized NMF
}
