% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarks.R
\name{runNMF}
\alias{runNMF}
\title{runNMF}
\usage{
runNMF(expression_matrix, k = 5, method = "lee", seed = 42)
}
\arguments{
\item{expression_matrix}{Matrix of gene expression data (samples as rows, genes as columns)}

\item{k}{A numeric variable representing the number of latent variables (aka, rank). Default is 5.}

\item{method}{A string representing the nmf algorithm to run. See nmf for options. Default is 'lee'.}

\item{seed}{Specifies the starting point or seeding method. See the NMF package for details.  Default is 42.}
}
\value{
A list containing the following components:
\item{H}{The matrix \code{H} obtained from the NMF calculation.}
\item{W}{The matrix \code{W} obtained from the NMF calculation.}
}
\description{
runNMF runs the version of NMF from the NMF package.
}
