% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{sampleLoadingsEvaluation}
\alias{sampleLoadingsEvaluation}
\title{sampleLoadingsEvaluation}
\usage{
sampleLoadingsEvaluation(W_test, test_metadata)
}
\arguments{
\item{W_test}{Matrix of projected W from test data}

\item{module_size}{Size of consecutive modules for DREAM simulations.}
}
\value{
top A list of the module score.
}
\description{
sampleLoadingsEvaluation is a function for identifying the top loading genes on each latent variable and aligning them with input data
}
