% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Similarity.R
\name{similarity_calc}
\alias{similarity_calc}
\title{similarity_calc}
\usage{
similarity_calc(input_matrix)
}
\arguments{
\item{input_matrix}{Input matrix of adjacency matrix or}
}
\description{
This function calculates the similarity matrix from an input matrix
}
