% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulateExpression}
\alias{simulateExpression}
\title{simualateExpression}
\usage{
simulateExpression(
  g,
  max_expression = 2000,
  iterations = 3,
  rc = NULL,
  select_nodes = NULL,
  perturbation_constant = 1.5,
  noise_constant = 0.3
)
}
\arguments{
\item{g}{igraph graph object}

\item{max_expression}{Numeric value for the maximum_gene expression. (Default is 2000)}

\item{iterations}{Numeric value for how many iterations to perform and find average (Default is 3)}

\item{rc}{Numeric vector of length 5 representing recaction and decay rates.}

\item{select_nodes}{Numeric vector of specific nodes to perturb}

\item{perturbation_constant}{Numberic value to increase initial values of perturbed nodes by. (Default is 1.5)}

\item{noise_constant}{Numeric value representing how much of the rest of the matrix to add noise to.}
}
\description{
simulateExpression is a function for simulating gene expression data from a GRN. It applies the functions outlined in the sgresR package vignette.
}
