% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulateMetadata}
\alias{simulateMetadata}
\title{simulateMetadata}
\usage{
simulateMetadata(group_size, group_labels, row_names)
}
\arguments{
\item{group_size}{Numeric vector for the size of each group}

\item{group_labels}{Character vector of group labels}

\item{row_names}{Character vector of rownames to assign the metadata}
}
\description{
generates simulated metadata based on cluster assignments
}
