% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_design.R
\name{split_data}
\alias{split_data}
\title{split_data}
\usage{
split_data(expression, metadata, training_size = 0.7)
}
\arguments{
\item{expression}{Data.frame of simulated gene expression data}

\item{metadata}{Data.frame of patient metadata}
}
\value{
train_expression Data.frame of training expression data

test_expression Data.frame of testing expression data

train_metadata Data.frame of training metadata

test_metadata Data.frame of testing metadata
}
\description{
split_data is a function for splitting the gene expression and patient metadata matrices into training and testing data.
}
