% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{categoricalPrediction}
\alias{categoricalPrediction}
\title{categoricalPredication}
\usage{
categoricalPrediction(W_test, test_metadata)
}
\arguments{
\item{W_test}{Matrix of projected W from test expression}

\item{test_metadata}{Matrix of test metadata}
}
\description{
Function for predicting categorical labels in test data
}
