% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{geneLoadingsEvaluation}
\alias{geneLoadingsEvaluation}
\title{geneLoadingsEvaluation}
\usage{
geneLoadingsEvaluation(H_train, clusters, aligned_clusters)
}
\arguments{
\item{H_train}{Matrix of trained H matrix}

\item{clusters}{Igraph object that is a list of cluster (aka, module) assignments}

\item{aligned_clusters}{The cluster labels aligned to the metadata group assignments}
}
\value{
top A list of the module score.
}
\description{
geneLoadingsEvaluation is a function for identifying the top loading genes on each latent variable and aligning them with input data
}
