% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{min_max_scale}
\alias{min_max_scale}
\title{min_max_scale}
\usage{
min_max_scale(x)
}
\arguments{
\item{x}{numerical vector}
}
\value{
scaled min max scaled numerical vector
}
\description{
min_max_scale is a function for scaling a vetor of values between 0 and 1
}
