% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DREAM_data.R, R/simulation_functions.R
\name{permuteNetwork}
\alias{permuteNetwork}
\title{permuteNetwork}
\usage{
permuteNetwork(g)

permuteNetwork(g)
}
\arguments{
\item{g}{igraph graph object}

\item{network}{Matrix for the adjacency to permute}
}
\value{
modified_matrix Matrix of the permuted network

permuted graph with the original degree distribution.
}
\description{
permuteNetwork is a function for permuting the network to test as a negative control

permuteNetwork is a function for permuting the network while maintaining the the degree distribution
n
}
