% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{randomNetwork}
\alias{randomNetwork}
\title{randomNetwork}
\usage{
randomNetwork(prior_graph, connected = TRUE)
}
\arguments{
\item{prior_graph}{Igraph object of the directed, weighted prior network to model  the simulated network on.}

\item{connected}{Boolean to determine if the random graph should be directed or not. (Default is TRUE)}
}
\value{
random graph with similar in and out degree
}
\description{
randomNetwork applies igraph functions for creating a random network
}
