% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulateExpression}
\alias{simulateExpression}
\title{simualateExpression}
\usage{
simulateExpression(g, iterations = 10, max_expression = 2000, num_samples = 5)
}
\arguments{
\item{g}{igraph graph object}

\item{iterations}{numeric indicating how many iterations to perform. (Default = 10)}

\item{max_expression}{Numeric value for the maximum_gene expression. (Default is 2000)}

\item{num_samples}{Numeric values for the number of samples to simulate. (Default is 5)}
}
\description{
simulateExpression is a function for simulating gene expression data from a GRN. It applies the functions outlined in the sgresR package vignette.
}
